Add files via upload
These are some maps I made, including pngs for with and without snow. I also included some code for the states of each of these maps. They are based on a coodinate system. Where the char is the x axis, and the int is the y axis. take a look when yall get the chance.