Add files via upload
State_gil.py is the same as state.py, only difference is that it includes a class for the plow. This class stores the homebase location, the current position, the available actions based on current position, the next tile based on current actions (for one tile step), Salt, and fuel levels.